+++ src/main/java/com/stemlink/skillmentor/services/StudentService.java
@@ -7,8 +7,8 @@ public interface StudentService {
 
     Student createNewStudent(Student student);
     List<Student> getAllStudents();
-    Student getStudentById(Integer id);
+    Student getStudentById(Long id);
     Student getStudentByStudentId(String studentId);
-    Student updateStudentById(Integer id, Student updatedStudent);
-    void deleteStudent(Integer id);
+    Student updateStudentById(Long id, Student updatedStudent);
+    void deleteStudent(Long id);
 }
